I added the SelectedView class which is essentially the panel but I did it in a JFrame since I didn't totally grasp the idea behind what you guys have done. There are still things I need to work on, like adding the images. We need to talk about the HP bar as well and how that will work.
I will line up the stats to the right later on and I will create space
to be able to draw the item images at the bottom as well. Let me know
what you guys think and if I should change anything.

I'll get working on the other JPanel now.